--- src/test/java/racingcar/RacingTest.java
@@ -1,4 +1,0 @@
-package racingcar;
-
-public class RacingTest {
-}
