+++ src/test/java/racingcar/domain/CarTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class MovingTest {
+public class CarTest {
 
     @Test
     @DisplayName("자동차가 움직인다")
@@ -31,62 +31,7 @@ public class MovingTest {
         assertThat(car.getPosition()).isEqualTo(0);
     }
 
-    @Test
-    @DisplayName("우승자가 한명일 때")
-    public void Only_One_Winner() {
-        // given
-        Car car1 = new Car("foo");
-        Car car2 = new Car("boo");
-        Car car3 = new Car("bar");
 
-        // when
-        car1.moveOrHold(true);
-        car1.moveOrHold(true);
-        car1.moveOrHold(true);
-
-        car2.moveOrHold(true);
-
-        car3.moveOrHold(true);
-        car3.moveOrHold(true);
-
-        CarRepository carRepository = new CarRepository();
-        GameController gameController = new GameController(carRepository);
-
-        carRepository.addCars(List.of(car1, car2, car3));
-        List<Car> winners = gameController.getWinners();
-
-        // then
-        assertThat(winners).containsExactly(car1);
-    }
-
-    @Test
-    @DisplayName("우승자가 여러일 때")
-    public void Two_Or_More_Winners() {
-        // given
-        Car car1 = new Car("foo");
-        Car car2 = new Car("boo");
-        Car car3 = new Car("bar");
-
-        // when
-        car1.moveOrHold(true);
-        car1.moveOrHold(true);
-        car1.moveOrHold(true);
-
-        car2.moveOrHold(true);
-
-        car3.moveOrHold(true);
-        car3.moveOrHold(true);
-        car3.moveOrHold(true);
-
-        CarRepository carRepository = new CarRepository();
-        GameController gameController = new GameController(carRepository);
-
-        carRepository.addCars(List.of(car1, car2, car3));
-        List<Car> winners = gameController.getWinners();
-
-        // then
-        assertThat(winners).containsExactly(car1, car3);
-    }
 
 
 }
